+++ src/Order/Events/OrderShippingAddressUpdated.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Order\Events;
+
+class OrderShippingAddressUpdated extends BaseOrderEvent
+{
+}